say whether upsert upped or serted
and start breaking subcommands out of base and into their own classes
