+++ lib/wavefront-cli/version.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-WF_CLI_VERSION = '5.1.0'
+WF_CLI_VERSION = '5.1.1'
